Modularized components, deleted unnecessary dependencies
