+++ client/components/Queuelist.js
@@ -1,12 +1,9 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import Queue from './queue';
 
 
 const QueueList = (props) => {
-  const list = props.queues.map(function(queue, ind) {
-    return <Queue key={ind} link={queue} />;
-  });
+  const list = props.queues.map((queue, ind) => <Queue key={ind} link={queue} />);
   return <div>{list}</div>;
 };
